Add initial simple parser (WIP)
Add a simple parser that takes care of sintax like `&Person.*`. The
statement can be parsed one argument at a time with `NextQueryArgument()`
until we run out of tokens or fully with the list of arguments returned by
`Parse()`.

`QueryArgument` saves the position of the _place holder_ in the
statement but does not hold a reference to it.